+++ common/src/main/java/com/feed_the_beast/ftbquests/client/FTBQuestsClientEventHandler.java
@@ -234,7 +234,7 @@ public class FTBQuestsClientEventHandler
 			}
 			else
 			{
-				for (int q : data.pinnedQuests)
+				for (long q : data.pinnedQuests)
 				{
 					Quest quest = file.getQuest(q);
