+++ common/src/main/java/com/feed_the_beast/ftbquests/quest/ServerQuestFile.java
@@ -120,7 +120,7 @@ public class ServerQuestFile extends QuestFile
 	}
 
 	@Override
-	public void deleteObject(int id)
+	public void deleteObject(long id)
 	{
 		QuestObjectBase object = getBase(id);
